Add subfolders and index files
