+++ apps/melding-form/src/app/(general)/Home.tsx
@@ -11,7 +11,7 @@ import { InvalidFormAlert } from '@meldingen/ui'
 import type { FormState } from '~/types'
 
 import { SystemErrorAlert } from './_components'
-import { getDocumentTitleOnError } from './_utils/validation/getDocumentTitleOnError'
+import { getDocumentTitleOnError } from './_utils/validation'
 
 const initialState: FormState = {}
