--- apps/melding-form/src/app/(general)/aanvullende-vragen/[classificationId]/[panelId]/_utils/navigationUtils.test.ts
@@ -1,169 +1,0 @@
-import type { PanelComponentsConditions } from './navigationUtils'
-
-import {
-  AFTER_ADDITIONAL_QUESTIONS_PATH,
-  getNextPanelPath,
-  getPreviousPanelPath,
-  refilterAnswersByKey,
-  shouldLinkToPanel,
-} from './navigationUtils'
-import { TOP_ANCHOR_ID } from '~/constants'
-
-const mockPanel = (key: string, conditions: PanelComponentsConditions['componentsConditions'] = []) => ({
-  componentsConditions: conditions,
-  key,
-})
-
-const doNotRenderWhenValueIsOne = { conditional: { eq: 'one', show: false, when: 'questionKey' }, key: 'key' }
-const renderWhenValueIsOne = { conditional: { eq: 'one', show: true, when: 'questionKey' }, key: 'key' }
-
-describe('shouldLinkToPanel', () => {
-  it('returns true when a component has no conditions', () => {
-    expect(shouldLinkToPanel(mockPanel('test', [{ conditional: null, key: 'not-used' }]), {})).toBe(true)
-  })
-
-  it('returns false when no components are rendered (show:false, condition met)', () => {
-    expect(shouldLinkToPanel(mockPanel('test', [doNotRenderWhenValueIsOne]), { questionKey: 'one' })).toBe(false)
-  })
-
-  it('returns true when at least one component is rendered', () => {
-    expect(
-      shouldLinkToPanel(mockPanel('test', [doNotRenderWhenValueIsOne, renderWhenValueIsOne]), { questionKey: 'one' }),
-    ).toBe(true)
-  })
-
-  it('returns true when show:false but condition is not met', () => {
-    expect(shouldLinkToPanel(mockPanel('test', [doNotRenderWhenValueIsOne]), { questionKey: 'two' })).toBe(true)
-  })
-
-  it('returns true when show:true and condition is met', () => {
-    expect(shouldLinkToPanel(mockPanel('test', [renderWhenValueIsOne]), { questionKey: 'one' })).toBe(true)
-  })
-
-  it('returns false when show:true but condition is not met', () => {
-    expect(shouldLinkToPanel(mockPanel('test', [renderWhenValueIsOne]), { questionKey: 'two' })).toBe(false)
-  })
-
-  it('returns true when answer is missing (cannot evaluate condition)', () => {
-    expect(shouldLinkToPanel(mockPanel('test', [doNotRenderWhenValueIsOne]), {})).toBe(true)
-  })
-
-  it('returns true when show:true, condition is met and answer is an array containing the value', () => {
-    expect(shouldLinkToPanel(mockPanel('test', [renderWhenValueIsOne]), { questionKey: ['one', 'two'] })).toBe(true)
-  })
-
-  it('returns false when show:true, condition is not met and answer is an array not containing the value', () => {
-    expect(shouldLinkToPanel(mockPanel('test', [renderWhenValueIsOne]), { questionKey: ['two', 'three'] })).toBe(false)
-  })
-
-  it('returns false when show:false, condition is met and answer is an array containing the value', () => {
-    expect(shouldLinkToPanel(mockPanel('test', [doNotRenderWhenValueIsOne]), { questionKey: ['one', 'two'] })).toBe(
-      false,
-    )
-  })
-})
-
-describe('getNextPanelPath', () => {
-  const panels = [mockPanel('panel-1'), mockPanel('panel-2', [doNotRenderWhenValueIsOne]), mockPanel('panel-3')]
-  const classificationId = 1
-  const currentPanelIndex = 0
-
-  it('returns the next panel that should be linked to', () => {
-    expect(getNextPanelPath(classificationId, currentPanelIndex, panels, {})).toBe(
-      `/aanvullende-vragen/1/panel-2#${TOP_ANCHOR_ID}`,
-    )
-  })
-
-  it('skips a hidden panel and returns the one after', () => {
-    expect(getNextPanelPath(classificationId, currentPanelIndex, panels, { questionKey: 'one' })).toBe(
-      `/aanvullende-vragen/1/panel-3#${TOP_ANCHOR_ID}`,
-    )
-  })
-
-  it('skips multiple consecutive hidden panels', () => {
-    const allHidden = [
-      mockPanel('panel-1'),
-      mockPanel('panel-2', [doNotRenderWhenValueIsOne]),
-      mockPanel('panel-3', [doNotRenderWhenValueIsOne]),
-      mockPanel('panel-4'),
-    ]
-    expect(getNextPanelPath(classificationId, currentPanelIndex, allHidden, { questionKey: 'one' })).toBe(
-      `/aanvullende-vragen/1/panel-4#${TOP_ANCHOR_ID}`,
-    )
-  })
-
-  it('returns AFTER_ADDITIONAL_QUESTIONS_PATH when all remaining panels are hidden', () => {
-    const allHiddenAfter = [mockPanel('panel-1'), mockPanel('panel-2', [doNotRenderWhenValueIsOne])]
-    expect(getNextPanelPath(classificationId, currentPanelIndex, allHiddenAfter, { questionKey: 'one' })).toBe(
-      AFTER_ADDITIONAL_QUESTIONS_PATH,
-    )
-  })
-
-  it('returns AFTER_ADDITIONAL_QUESTIONS_PATH when already on the last panel', () => {
-    expect(getNextPanelPath(classificationId, 2, panels, {})).toBe(AFTER_ADDITIONAL_QUESTIONS_PATH)
-  })
-})
-
-const beforePath = '/before'
-
-describe('getPreviousPanelPath', () => {
-  const panels = [mockPanel('panel-1'), mockPanel('panel-2', [doNotRenderWhenValueIsOne]), mockPanel('panel-3')]
-  const classificationId = 1
-
-  it('returns the previous panel that should be linked to', () => {
-    expect(getPreviousPanelPath(classificationId, 2, panels, {}, beforePath)).toBe(
-      `/aanvullende-vragen/1/panel-2#${TOP_ANCHOR_ID}`,
-    )
-  })
-
-  it('skips a hidden panel and returns the one before', () => {
-    expect(getPreviousPanelPath(classificationId, 2, panels, { questionKey: 'one' }, beforePath)).toBe(
-      `/aanvullende-vragen/1/panel-1#${TOP_ANCHOR_ID}`,
-    )
-  })
-
-  it('skips multiple consecutive hidden panels', () => {
-    const allHidden = [
-      mockPanel('panel-1'),
-      mockPanel('panel-2', [doNotRenderWhenValueIsOne]),
-      mockPanel('panel-3', [doNotRenderWhenValueIsOne]),
-      mockPanel('panel-4'),
-    ]
-    expect(getPreviousPanelPath(classificationId, 3, allHidden, { questionKey: 'one' }, beforePath)).toBe(
-      `/aanvullende-vragen/1/panel-1#${TOP_ANCHOR_ID}`,
-    )
-  })
-
-  it('returns beforePath when all preceding panels are hidden', () => {
-    const allHiddenBefore = [mockPanel('panel-1', [doNotRenderWhenValueIsOne]), mockPanel('panel-2')]
-    expect(getPreviousPanelPath(classificationId, 1, allHiddenBefore, { questionKey: 'one' }, beforePath)).toBe(
-      beforePath,
-    )
-  })
-
-  it('returns beforePath when already on the first panel', () => {
-    expect(getPreviousPanelPath(classificationId, 0, panels, {}, beforePath)).toBe(beforePath)
-  })
-})
-
-describe('refilterAnswersByKey', () => {
-  const panels: PanelComponentsConditions[] = [
-    mockPanel('panel-1', [{ key: 'key1' }, { conditional: { eq: 'one', show: true, when: 'key1' }, key: 'key2' }]),
-    mockPanel('panel-2', [{ key: 'key3' }]),
-  ]
-
-  it('returns only answers for components that should be rendered', () => {
-    const answers = { key1: 'two', key2: 'answer2', key3: 'answer3' }
-    expect(refilterAnswersByKey(panels, answers)).toEqual({ key1: 'two', key3: 'answer3' })
-  })
-
-  it('returns an empty object when answers contain no matching keys', () => {
-    const answers = { key4: 'answer4', key5: 'answer5', key6: 'answer6' }
-    expect(refilterAnswersByKey(panels, answers)).toEqual({})
-  })
-
-  it('returns all answers when all components should be rendered', () => {
-    const answers = { key1: 'one', key2: 'answer2', key3: 'answer3' }
-    expect(refilterAnswersByKey(panels, answers)).toEqual({ key1: 'one', key2: 'answer2', key3: 'answer3' })
-  })
-})
