+++ apps/melding-form/src/app/(general)/aanvullende-vragen/[classificationId]/[panelId]/_utils/processEntries/index.ts
@@ -1,0 +1,3 @@
+export { categorizeFormEntries } from './categorizeFormEntries'
+export { mergeCheckboxAnswers } from './mergeCheckboxAnswers'
+export { mergeUnknownTimeAnswers } from './mergeUnknownTimeAnswers'
