+++ apps/melding-form/src/app/(general)/actions.ts
@@ -6,7 +6,7 @@ import { redirect } from 'next/navigation'
 import { patchMeldingByMeldingIdMelder, postMelding } from '@meldingen/api-client'
 
 import { resolveClassificationRedirect } from '../_utils/resolveClassificationRedirect'
-import { hasValidationErrors } from './_utils/validation/hasValidationErrors'
+import { hasValidationErrors } from './_utils/validation'
 import { handleApiError } from '~/app/_utils/handleApiError'
 import { COOKIES } from '~/constants'
