+++ apps/melding-form/src/app/(general)/samenvatting/Summary.tsx
@@ -11,7 +11,7 @@ import { Link, SubmitButton, SummaryList, UnorderedList } from '@meldingen/ui'
 import type { FormState } from '~/types'
 
 import { SystemErrorAlert } from '../_components'
-import { getDocumentTitleOnError } from '../_utils/validation/getDocumentTitleOnError'
+import { getDocumentTitleOnError } from '../_utils/validation'
 import { BackLink } from '../../_components'
 import { AttachmentImage } from './_components/AttachmentImage'
 import { TOP_ANCHOR_ID } from '~/constants'
